ref(front): Sort data types and interfaces
