+++ src/WebApp/smartbusiness.web/src/api/apiReadConnector.ts
@@ -1,5 +1,5 @@
 import { axiosRead } from "./axiosInstance";
-import type { Transaction } from "../models/index";
+import type { Transaction } from "../models/transaction";
 
 const apiReadConnector = {
