+++ src/WebApp/smartbusiness.web/src/api/apiWriteConnector.ts
@@ -1,5 +1,5 @@
 import { axiosWrite } from "./axiosInstance";
-import type { Transaction } from "../models/index";
+import type { Transaction } from "../models/transaction";
 
 const apiWriteConnector = {
