+++ src/WebApp/smartbusiness.web/src/components/Dashboard/Company/DashboardHomeSection.tsx
@@ -5,8 +5,7 @@ export const DashboardHomeSection: React.FC = () => {
   const { company } = useCompany();
   return (
     <>
-      {/* Statystyki, karty, info o firmie, powitanie itp. */}
-      {/* ...przeniesiona zawartość z dotychczasowego DashboardPage... */}
+      {/* Header Section */}
       <div className="mb-8">
         <h1 className="mb-2 text-3xl font-bold text-gray-900 dark:text-gray-100">
           Dashboard
@@ -15,9 +14,10 @@ export const DashboardHomeSection: React.FC = () => {
           Zarządzaj swoim kontem i aplikacją
         </p>
       </div>
+
       {/* Stats Cards */}
       <div className="mb-8 grid grid-cols-1 gap-6 md:grid-cols-3">
-        {/* ...karty statystyk... */}
+        {/* Some content in the future */}
       </div>
       {/* Company Info Card */}
       <div className="rounded-xl border border-gray-200 bg-gray-100 p-6 shadow-xl dark:border-gray-900 dark:bg-gray-800">
@@ -49,6 +49,7 @@ export const DashboardHomeSection: React.FC = () => {
           </div>
         )}
       </div>
+
       {/* Welcome Message */}
       <div className="mt-8 rounded-lg bg-gradient-to-r from-indigo-500 to-purple-600 p-6 text-white shadow">
         <h2 className="mb-2 text-2xl font-bold">Witaj w SecureApp!</h2>
