+++ src/WebApp/smartbusiness.web/src/models/authErrors.ts
@@ -9,4 +9,3 @@ export type ApiResponseError = {
   detail: string;
   errors?: ApiResponseValidationError[] | null;
 };
-
