+++ src/WebApp/smartbusiness.web/src/models/transaction.ts
@@ -1,0 +1,12 @@
+export interface Transaction {
+  id: string;
+  companyId: string;
+  userId: string;
+  productId: string;
+  quantity: number;
+  totalAmount: number;
+  tax: number;
+  totalAmountMinusTax: number;
+  createdAt: string; // ISO string (DateTime)
+}
+
